chore: Add TimePeriod addition functionality
This commit adds a new function to the TimePeriod class that allows two TimePeriod objects to be summed together. The function `__add__` takes another TimePeriod object as input and returns a new TimePeriod object representing the sum of the two time periods. This functionality will be useful for calculating cumulative experience years in the future.